move color related struct & functions to colors.rs
