+++ src/config/mod.rs
@@ -1,17 +1,16 @@
 mod cloud_config;
-mod configurable_static;
+mod colors;
 mod configuration;
 mod gradient;
 mod keybindings;
+mod oncelock_static;
 
 pub use cloud_config::cloud_config;
-pub use configurable_static::{
-    set_configurable_static, COLORER, FILE_ATTRS, GRADIENT_NORMAL_FILE, MENU_ATTRS, MONOKAI_THEME,
+pub use colors::{extension_color, str_to_tuikit, ColorG, NormalFileColorer};
+pub use configuration::{load_config, read_normal_file_colorer, Config, FileAttr, MenuAttrs};
+pub use gradient::Gradient;
+pub use keybindings::Bindings;
+pub use oncelock_static::{
+    set_configurable_static, ARRAY_GRADIENT, COLORER, FILE_ATTRS, MENU_ATTRS, MONOKAI_THEME,
     START_FOLDER,
 };
-pub use configuration::{
-    load_color_from_config, load_config, parse_text_triplet, read_normal_file_colorer, Config,
-    FileAttr, MenuAttrs,
-};
-pub use gradient::{extension_color, ColorG, Gradient, NormalFileColorer};
-pub use keybindings::Bindings;
